requirements.txt
pytest~=8.3.5
allure-pytest
requests~=2.32.3
allure-python-commons~=2.14.2